fix: usage types

+++ src/routes/dashboard/usage.rs
@@ -4,7 +4,7 @@ pub async fn usage_handler(
     Extension(claims): Extension<Claims>,
     State(state): State<AppState>,
 ) -> Result<Json<UsageResponse>, DashboardError> {
-    let result: Option<i32> =
+    let result: Option<i64> =
         sqlx::query_scalar("SELECT SUM(requests) FROM databases WHERE owner_id = $1")
             .bind(&claims.user_id)
             .fetch_one(&state.pool)
